Abstract out shared logic between tutorials and code examples pages
