+++ src/pages/videos.tsx
@@ -4,7 +4,7 @@ import { PillCategory } from '../types/pill-category';
 
 import { ContentTypePageProps } from '../page-templates/content-type/types';
 import { getFilters } from '../page-templates/content-type/utils';
-import getL1Content from '../mockdata/get-l1-content';
+import { getAllContentItems } from '../service/get-all-content';
 
 const VideosPage: NextPage<ContentTypePageProps> = props => {
     return <ContentTypePage {...props} />;
@@ -15,7 +15,12 @@ export const getStaticProps: GetStaticProps = async () => {
     // Pop contentTypeItems out of here becasue we don't filter by it for these pages.
     const { contentTypeItems, ...filters } = await getFilters(contentType);
 
-    const { featured } = getL1Content();
+    // TODO: When the featured collection is populated, we will hit that and filter by content type.
+    const content = await getAllContentItems();
+    const featured = content
+        .filter(item => item.category === contentType)
+        .sort((a, b) => b.contentDate.localeCompare(a.contentDate))
+        .slice(0, 3);
 
     return {
         props: { contentType, ...filters, featured },
